refactor(bfabric_scripts): de-duplicate auth login tests
Extract the repeated OAuth token/session boilerplate and the thrice-copied
BFABRICPY_CONFIG_ENV cleanup into a login-dir conftest fixture set, and drop
the redundant repeated ask()-is-Any comments from the interactive wrappers.
No behavior change; login + interactive suites (83 tests) green.
