+++ tests/bfabric_scripts/cli/login/test_cmd_login_device_code.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-import time
-
 import yaml
 
 from bfabric._oauth._constants import DEFAULT_OAUTH_SCOPE
@@ -9,21 +7,9 @@ from bfabric_scripts.cli.login.device_code import cmd_login_device_code
 
 
 class TestCmdLoginDeviceCode:
-    def test_writes_config_and_caches_token(self, tmp_path, mocker):
+    def test_writes_config_and_caches_token(self, tmp_path, mocker, oauth_token, oauth_session):
         config_file = tmp_path / "config.yml"
-        token = {
-            "access_token": "jwt123",
-            "refresh_token": "rt456",
-            "token_type": "Bearer",
-            "expires_at": time.time() + 3600,
-        }
-
-        mock_session = mocker.MagicMock()
-        mock_session.token = None
-        mock_session.metadata = {"token_endpoint": "https://example.com/bfabric/rest/oauth/token"}
-
-        mock_dc = mocker.patch("bfabric_scripts.cli.login.device_code.device_code_login", return_value=token)
-        mocker.patch("bfabric._oauth.credential_provider.OAuth2Session", return_value=mock_session)
+        mock_dc = mocker.patch("bfabric_scripts.cli.login.device_code.device_code_login", return_value=oauth_token)
         cmd_login_device_code(
             base_url="https://example.com/bfabric",
             client_id="test-client",
@@ -37,20 +23,9 @@ class TestCmdLoginDeviceCode:
         assert data["PROD"]["client_id"] == "test-client"
         assert data["PROD"]["base_url"] == "https://example.com/bfabric"
 
-    def test_set_default_false_does_not_set_default(self, tmp_path, mocker):
+    def test_set_default_false_does_not_set_default(self, tmp_path, mocker, oauth_token, oauth_session):
         config_file = tmp_path / "config.yml"
-        token = {
-            "access_token": "jwt123",
-            "refresh_token": "rt456",
-            "token_type": "Bearer",
-            "expires_at": time.time() + 3600,
-        }
-        mock_session = mocker.MagicMock()
-        mock_session.token = None
-        mock_session.metadata = {"token_endpoint": "https://example.com/bfabric/rest/oauth/token"}
-
-        mocker.patch("bfabric_scripts.cli.login.device_code.device_code_login", return_value=token)
-        mocker.patch("bfabric._oauth.credential_provider.OAuth2Session", return_value=mock_session)
+        mocker.patch("bfabric_scripts.cli.login.device_code.device_code_login", return_value=oauth_token)
         cmd_login_device_code(
             base_url="https://example.com/bfabric",
             client_id="test-client",
@@ -63,19 +38,9 @@ class TestCmdLoginDeviceCode:
         assert "default_config" not in data["GENERAL"]
         assert data["PROD"]["auth_method"] == "oauth"
 
-    def test_scope_preset_is_expanded(self, tmp_path, mocker):
+    def test_scope_preset_is_expanded(self, tmp_path, mocker, oauth_token, oauth_session):
         config_file = tmp_path / "config.yml"
-        token = {
-            "access_token": "jwt123",
-            "refresh_token": "rt456",
-            "token_type": "Bearer",
-            "expires_at": time.time() + 3600,
-        }
-        mock_session = mocker.MagicMock()
-        mock_session.token = None
-        mock_session.metadata = {"token_endpoint": "https://example.com/bfabric/rest/oauth/token"}
-        mock_dc = mocker.patch("bfabric_scripts.cli.login.device_code.device_code_login", return_value=token)
-        mocker.patch("bfabric._oauth.credential_provider.OAuth2Session", return_value=mock_session)
+        mock_dc = mocker.patch("bfabric_scripts.cli.login.device_code.device_code_login", return_value=oauth_token)
         cmd_login_device_code(
             base_url="https://example.com/bfabric",
             client_id="test-client",
