+++ tests/bfabric_scripts/cli/login/test_cmd_login_pkce.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-import time
-
 import yaml
 
 from bfabric._oauth._constants import DEFAULT_OAUTH_SCOPE
@@ -9,21 +7,9 @@ from bfabric_scripts.cli.login.pkce import cmd_login_pkce
 
 
 class TestCmdLoginPkce:
-    def test_writes_config_and_caches_token(self, tmp_path, mocker):
+    def test_writes_config_and_caches_token(self, tmp_path, mocker, oauth_token, oauth_session):
         config_file = tmp_path / "config.yml"
-        token = {
-            "access_token": "jwt123",
-            "refresh_token": "rt456",
-            "token_type": "Bearer",
-            "expires_at": time.time() + 3600,
-        }
-
-        mock_session = mocker.MagicMock()
-        mock_session.token = None
-        mock_session.metadata = {"token_endpoint": "https://example.com/bfabric/rest/oauth/token"}
-
-        mock_pkce = mocker.patch("bfabric_scripts.cli.login.pkce.pkce_login", return_value=token)
-        mocker.patch("bfabric._oauth.credential_provider.OAuth2Session", return_value=mock_session)
+        mock_pkce = mocker.patch("bfabric_scripts.cli.login.pkce.pkce_login", return_value=oauth_token)
         cmd_login_pkce(
             base_url="https://example.com/bfabric",
             client_id="test-client",
@@ -37,20 +23,9 @@ class TestCmdLoginPkce:
         assert data["PROD"]["client_id"] == "test-client"
         assert data["PROD"]["base_url"] == "https://example.com/bfabric"
 
-    def test_set_default_false_does_not_set_default(self, tmp_path, mocker):
+    def test_set_default_false_does_not_set_default(self, tmp_path, mocker, oauth_token, oauth_session):
         config_file = tmp_path / "config.yml"
-        token = {
-            "access_token": "jwt123",
-            "refresh_token": "rt456",
-            "token_type": "Bearer",
-            "expires_at": time.time() + 3600,
-        }
-        mock_session = mocker.MagicMock()
-        mock_session.token = None
-        mock_session.metadata = {"token_endpoint": "https://example.com/bfabric/rest/oauth/token"}
-
-        mocker.patch("bfabric_scripts.cli.login.pkce.pkce_login", return_value=token)
-        mocker.patch("bfabric._oauth.credential_provider.OAuth2Session", return_value=mock_session)
+        mocker.patch("bfabric_scripts.cli.login.pkce.pkce_login", return_value=oauth_token)
         cmd_login_pkce(
             base_url="https://example.com/bfabric",
             client_id="test-client",
@@ -63,19 +38,9 @@ class TestCmdLoginPkce:
         assert "default_config" not in data["GENERAL"]
         assert data["PROD"]["auth_method"] == "oauth"
 
-    def test_prompts_for_default_when_omitted(self, tmp_path, mocker):
+    def test_prompts_for_default_when_omitted(self, tmp_path, mocker, oauth_token, oauth_session):
         config_file = tmp_path / "config.yml"
-        token = {
-            "access_token": "jwt123",
-            "refresh_token": "rt456",
-            "token_type": "Bearer",
-            "expires_at": time.time() + 3600,
-        }
-        mock_session = mocker.MagicMock()
-        mock_session.token = None
-        mock_session.metadata = {"token_endpoint": "https://example.com/bfabric/rest/oauth/token"}
-        mocker.patch("bfabric_scripts.cli.login.pkce.pkce_login", return_value=token)
-        mocker.patch("bfabric._oauth.credential_provider.OAuth2Session", return_value=mock_session)
+        mocker.patch("bfabric_scripts.cli.login.pkce.pkce_login", return_value=oauth_token)
         # No --set-default given: in a terminal the user is asked; here they decline.
         mocker.patch("bfabric_scripts.cli.login._common.is_interactive", return_value=True)
         confirm = mocker.patch("bfabric_scripts.cli.login._common.confirm", return_value=False)
@@ -92,19 +57,9 @@ class TestCmdLoginPkce:
         assert "default_config" not in data["GENERAL"]
         assert data["PROD"]["auth_method"] == "oauth"
 
-    def test_scope_preset_is_expanded(self, tmp_path, mocker):
+    def test_scope_preset_is_expanded(self, tmp_path, mocker, oauth_token, oauth_session):
         config_file = tmp_path / "config.yml"
-        token = {
-            "access_token": "jwt123",
-            "refresh_token": "rt456",
-            "token_type": "Bearer",
-            "expires_at": time.time() + 3600,
-        }
-        mock_session = mocker.MagicMock()
-        mock_session.token = None
-        mock_session.metadata = {"token_endpoint": "https://example.com/bfabric/rest/oauth/token"}
-        mock_pkce = mocker.patch("bfabric_scripts.cli.login.pkce.pkce_login", return_value=token)
-        mocker.patch("bfabric._oauth.credential_provider.OAuth2Session", return_value=mock_session)
+        mock_pkce = mocker.patch("bfabric_scripts.cli.login.pkce.pkce_login", return_value=oauth_token)
         cmd_login_pkce(
             base_url="https://example.com/bfabric",
             client_id="test-client",
@@ -115,7 +70,7 @@ class TestCmdLoginPkce:
         # The preset name expands to the real scope string requested from the OAuth flow.
         assert mock_pkce.call_args.kwargs["scope"] == f"{DEFAULT_OAUTH_SCOPE} tus"
 
-    def test_config_env_omitted_falls_back_to_current_default(self, tmp_path, mocker):
+    def test_config_env_omitted_falls_back_to_current_default(self, tmp_path, mocker, oauth_token, oauth_session):
         config_file = tmp_path / "config.yml"
         config_file.write_text(
             yaml.dump(
@@ -125,17 +80,7 @@ class TestCmdLoginPkce:
                 }
             )
         )
-        token = {
-            "access_token": "jwt123",
-            "refresh_token": "rt456",
-            "token_type": "Bearer",
-            "expires_at": time.time() + 3600,
-        }
-        mock_session = mocker.MagicMock()
-        mock_session.token = None
-        mock_session.metadata = {"token_endpoint": "https://example.com/bfabric/rest/oauth/token"}
-        mocker.patch("bfabric_scripts.cli.login.pkce.pkce_login", return_value=token)
-        mocker.patch("bfabric._oauth.credential_provider.OAuth2Session", return_value=mock_session)
+        mocker.patch("bfabric_scripts.cli.login.pkce.pkce_login", return_value=oauth_token)
         # No TTY (pytest) and no --config-env => reuse the current default env, not PRODUCTION.
         mocker.patch("bfabric_scripts.cli.login._common.is_interactive", return_value=False)
         cmd_login_pkce(base_url="https://example.com/bfabric", client_id="c", config_file=config_file)
